move cli files
